Remove classes that just wrap NLTK
Removing these simplifies the codebase and makes it easier for
developers who already know NLTK to apply their knowledge.
